Remove unnecessary include lines from header files.
This is a step in making the public interface of a module minimal.

Create a tool to run a modify-build-loop that removes as many include
lines as possible. Start off with

for f in src/*.c; do
    hf=${f/.c/.h}
    if ! [ -f "$hf" ]; then
       continue
    fi
    ./tools/minhdrmake "$hf" make -j2 "${f/.c/.o}" || break
done

and fix mistakes and add include lines in other source files as needed.