Remove DiagFlag, unexport NBifs, start_diagram and edit_start.
The previous code used a complex linked-list creation structure. We
simplify, but allow bifd to be NULL.

save_diagram and load_diagram seemed to have an off-by-one issue
when only one diagram was available. This breaks the file format,
but only for cases where the file has no diagrams anyway.