Move options defaults to first. Remove support for options file.
Start deprecating notAlreadySet by moving setting of built-in defaults
to happen first. Loading the options file from the ODE file proves
tricky since it has a fixed format, and the parser is fragile (whether a
line is expected depends on notAlreadySet). So, support for default.opt
is removed. Use xpprc or @ instead.

Previously, for a file on cmdline, the flow was

  check_for_quiet
  do_cmdline

  set_X_vals()
  check_for_xpprc()
  set_internopts_xpprc_and_comline()
  set_all_vals()
  set_internopts((nil))
  read_defaults

  set_internopts(0x1d4cab0)

  set_X_vals()
  check_for_xpprc()
  set_internopts_xpprc_and_comline()
  set_all_vals()
  set_internopts((nil))
  read_defaults

In the new code, the flow is

  loadeqn_init_options()
  do_cmdline
  check_for_xpprc()
  set_internopts_xpprc_and_comline()

  set_internopts(0x2ca1ac0)
  check_for_xpprc()
  set_internopts_xpprc_and_comline()
  loadeqn_setup_all()